requirements.txt
openai
rich
streamlit 
python-dotenv